requirements.txt
streamlit==1.9.0
requests==2.27.1
s3fs==0.4.2
Pillow==9.1.1
boto3==1.21.32
botocore==1.24.32
sqlalchemy==1.4.32